requirements.txt
langchain==0.2.6
langchain-openai==0.1.13
langchain-community==0.2.6
openai==1.35.7
pydantic==2.7.4
python-dotenv==1.0.1
rich==13.7.1
pytest==8.2.2